implement integration event listening
