+++ src/message_handlers/authenticated_plugin_user.ts
@@ -1,7 +1,5 @@
 import { authenticationCompleteHandler } from "../DataController";
 
 export async function handleAuthenticatedPluginUser(user: any) {
-  console.debug("[CodeTime] Received authenticated plugin user message", user);
-
   authenticationCompleteHandler(user);
 }
